requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
click==7.1.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
paho-mqtt==1.5.0
pathspec==0.8.0
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.0
regex==2020.10.28
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
websockets==8.1
wrapt==1.12.1
